:star2: feat: load store via backend
