+++ src/backend/firebase/index.js
@@ -1,0 +1,7 @@
+import init from './init'
+import store from './store'
+
+export default {
+  init,
+  store
+}
